feat: station services foundation (Spec 006 Phase 1-2)
- Create StationServices feature module with Data/Systems/Views/Tests
- Add state types: StationServicesState, StationStorageState, RefiningJobState,
  RefiningOutputConfig, MaterialOutput, RefiningJobStatus to Core.State
- Add IStationServicesAction marker interface and all 11 action records
- Add 6 event structs for UI feedback
- Add ScriptableObjects: StationServicesConfig, GameServicesConfig,
  RawMaterialDefinition, StationServicesConfigMap
- Implement StationStorageReducer and StationServicesReducer (pure static)
- Implement RepairMath with ceiling rounding
- Update CompositeReducer with cross-cutting handlers (transfer, repair)
- Initialize per-station storage in CreateDefaultGameState
- Migrate BaseValue float→int in OreDefinition
- Move station services menu stub from Docking to StationServices
- Replace RefiningState with StationServicesState in GameLoopState
- 375 tests pass, 0 regressions

+++ Assets/Features/Mining/Data/OreDefinition.cs
@@ -22,8 +22,8 @@ namespace VoidHarvest.Features.Mining.Data
         /// <summary>Inventory/UI icon. Nullable — stored for future use.</summary>
         public Sprite Icon;
 
-        /// <summary>Base market value per unit. Stored for future economy integration.</summary>
-        public float BaseValue;
+        /// <summary>Base market value per unit in integer credits.</summary>
+        public int BaseValue;
 
         /// <summary>Flavor text for tooltips. Stored for future use.</summary>
         [TextArea]
